requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
discord.py @ git+https://github.com/Rapptz/discord.py@ab287e71b8961cbe245e60ad220b7d0be9fd0b8a
frozenlist==1.3.3
gTTS==2.3.1
idna==3.4
multidict==6.0.4
python-dotenv==1.0.0
requests==2.28.2
urllib3==1.26.15
yarl==1.8.2
